// packages/api-client/src/generated-types/InsightsQueryParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs).
// Relative imports converted to absolute imports by generate-types recipe. DO NOT EDIT.
import type { InsightSeverity } from "@/generated-types/InsightSeverity";
import type { InsightType } from "@/generated-types/InsightType";
import type { InsightsSortBy } from "@/generated-types/InsightsSortBy";
import type { SortOrder } from "@/generated-types/SortOrder";

export type InsightsQueryParams = {
  page?: number;
  limit?: number;
  search?: string;
  severity?: InsightSeverity;
  insight_type?: InsightType;
  start?: bigint;
  end?: bigint;
  sort_by?: InsightsSortBy;
  sort_order?: SortOrder;
};